Add base class ParserServer.java and refactorize CalciteParserServer.java
