+++ parser/remote/grpc/pom.xml
@@ -16,15 +16,15 @@
 -->
 
 <project xmlns="http://maven.apache.org/POM/4.0.0"
-xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+<modelVersion>4.0.0</modelVersion>
 <parent>
   <groupId>org.sqlflow</groupId>
   <artifactId>remote-parsers</artifactId>
   <version>0.1-SNAPSHOT</version>
   <relativePath>../pom.xml</relativePath>
 </parent>
-<modelVersion>4.0.0</modelVersion>
 <artifactId>parser-grpc</artifactId>
 
 <build>
